pyproject.toml
[project]
name = "wiki-gap"
version = "0.1.0"
description = "Wikipedia 日英ギャップ検出ツール (医学系)"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "SPARQLWrapper>=2.0.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
